Move FieldConfigEditor to core
