+++ packages/grafana-data/src/types/fieldOverrides.ts
@@ -1,12 +1,7 @@
 import { ComponentType } from 'react';
-import { MatcherConfig, FieldConfig, Field, DataFrame } from '../types';
+import { MatcherConfig, FieldConfig, Field, DataFrame, VariableSuggestion, VariableSuggestionsScope } from '../types';
 import { Registry, RegistryItem } from '../utils';
 import { InterpolateFunction } from './panel';
-import { VariableSuggestion } from '../links/variableSuggestions';
-
-export enum VariableSuggestionsScope {
-  Values = 'values',
-}
 
 export interface DynamicConfigValue {
   prop: string;
