+++ packages/grafana-ui/src/components/DataLinks/DataLinkInput.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useMemo, useContext, useRef, RefObject, memo, useEffect } from 'react';
 import usePrevious from 'react-use/lib/usePrevious';
-import { VariableSuggestion, VariableOrigin, DataLinkSuggestions } from './DataLinkSuggestions';
+import { DataLinkSuggestions } from './DataLinkSuggestions';
 import { ThemeContext, DataLinkBuiltInVars, makeValue } from '../../index';
 import { SelectionReference } from './SelectionReference';
 import { Portal } from '../index';
@@ -14,7 +14,7 @@ import { css, cx } from 'emotion';
 import { SlatePrism } from '../../slate-plugins';
 import { SCHEMA } from '../../utils/slate';
 import { stylesFactory } from '../../themes';
-import { GrafanaTheme } from '@grafana/data';
+import { GrafanaTheme, VariableSuggestion, VariableOrigin } from '@grafana/data';
 
 const modulo = (a: number, n: number) => a - n * Math.floor(a / n);
