+++ public/app/plugins/datasource/elasticsearch/configuration/DataLinks.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { css } from 'emotion';
-import { Button, DataLinkBuiltInVars, stylesFactory, useTheme, VariableOrigin } from '@grafana/ui';
-import { GrafanaTheme } from '@grafana/data';
+import { Button, DataLinkBuiltInVars, stylesFactory, useTheme } from '@grafana/ui';
+import { GrafanaTheme, VariableOrigin } from '@grafana/data';
 import { DataLinkConfig } from '../types';
 import { DataLink } from './DataLink';
